Lift and refactor `editorialPalette`
- Moved AR's `editorialPalette` to `common-rendering`
- Merged with `common-rendering`'s `editorialPalette`
- Removed terms "primary" and "inverse"
- Renamed `border.primary` to `border.articleLink` to match DCR
- Updated imports
