+++ apps-rendering/src/components/headline.tsx
@@ -8,7 +8,7 @@ import { neutral, remSpace } from '@guardian/src-foundations';
 import { between, from } from '@guardian/src-foundations/mq';
 import { headline, textSans } from '@guardian/src-foundations/typography';
 import StarRating from 'components/starRating';
-import { border } from 'editorialPalette';
+import { border } from '@guardian/common-rendering/src/editorialPalette';
 import { headlineBackgroundColour, headlineTextColour } from 'editorialStyles';
 import type { Item } from 'item';
 import type { ReactElement } from 'react';
@@ -67,7 +67,7 @@ const analysisStyles = (format: ArticleFormat): SerializedStyles => css`
 	${headline.medium({ lineHeight: 'regular', fontWeight: 'light' })}
 
 	span {
-		box-shadow: inset 0 -0.025rem ${border.primary(format)};
+		box-shadow: inset 0 -0.025rem ${border.articleLink(format)};
 		padding-bottom: 0.2rem;
 
 		${darkModeCss`
